requirements.txt
django>=2.0,<2.1
djangorestframework>=3.7,<3.8
Markdown>=2.6,<2.7
